Homepage is complete!
All the courses/clusters and units the user is enrolled into are displayed on the main screen based on the fact that a student must be full-time. If the user is part-time, we must filter either the clusters or units to conform to that need. The HomepageHelper contains quite a few methods to retrieve courses, clusters and units if necessary, to reference it, just follow the same syntaxt from Application_pages_controller and HomepageHelper.